+++ NEWS.md
@@ -2,6 +2,12 @@
 
 * Developmental version of package.
 
+* Added unit tests.
+
+* Patched a small bug with prevented univariate random-intercept models for being fitted.
+
+* Added a check that any initial covariance matrix given is positive-definite.
+
 * LICENSE upgraded to GPL-3 to be compatible with `joineR` v1.1.0.
 
 * Package now `Depends` on `survival` and `nlme` rather than `Imports` to allow `require()` statements to be removed from code.